Add files via upload
A bar chart or graph is used to represent category wise data of a dataset. The length of the bars is proportional to the value that each category represents. This gives instant insight into the data pictorially. In Tableau, we can create several variants like a simple vertical bar graph (also known as a line graph), segmented bar graph, stacked bar graph and side by side bar chart.